// src/interfaces/gameLogStats.interface.ts
export default interface IGAME_LOG_STATS {
  date?: string;
  opp?: string;
  min?: string;
  pts?: string;
  fg?: string;
  threeFg?: string;
  ft?: string;
  offReb?: string;
  defReb?: string;
  reb?: string;
  ast?: string;
  stl?: string;
  blk?: string;
  pf?: string;
  to?: string;
  plusMinus?: string;
}